multiple changes
changed current measure to integers to limit the size of the status message better
added length checks on all commands.  added some parameter checks on individual commands.  It would probably be good to make sure conversions from string to integer return a number.
fixed error in implementation of adjust power command.  I had it as a 3 byte value, when it should have been 1 byte, with power levels 0-10.  I had to also fix this in the GUI.
Still Testing.

+++ e2ecommands_metro/efuse.h
@@ -47,14 +47,14 @@ class Efuse
     Efuse(int I_Monitor_pin, int OC5V_pin, int Efuse_reset_pin);
 
     void begin();
-    float measure_current();
+    int measure_current();
     int overcurrent(bool transmit);
     bool fault();
     void reset();
     void set_OC_threshold(float threshold, bool transmit);
     void set_imon_slope(float imon_slope);
     void set_imon_intercept(float imon_intercept);
-    float get_max_current();
+    int get_max_current();
     void clear_max_current();
 
   private:
